refactor: split link checker modules
